+++ src/qml/dialogs/WaypointDescription.qml
@@ -42,7 +42,8 @@ CenteringDialog {
         // Delete old text items
         let childCount = co.children.length;
         // Iterate through the children in reverse order
-        for (let i = childCount - 1; i >= 0; i--) {
+        var i
+        for (i = childCount - 1; i >= 0; i--) {
             // Check if the child is a valid QML item
             if (co.children[i] instanceof QtObject) {
                     // Destroy the child item
@@ -67,7 +68,7 @@ CenteringDialog {
 
         // Create airspace description items
         var asl = GeoMapProvider.airspaces(waypoint.coordinate)
-        for (var i in asl)
+        for (i in asl)
             airspaceDelegate.createObject(co, {airspace: asl[i]});
 
         satButtonDelegate.createObject(co, {});
@@ -484,17 +485,19 @@ CenteringDialog {
 
                     onTriggered: {
                         PlatformAdaptor.vibrateBrief()
-                        if (Navigator.flightRoute.size > 0)
+                        if (Navigator.flightRoute.size > 0) {
+                            addMenu.close()
                             overwriteDialog.open()
+                        }
                         else {
                             Navigator.flightRoute.clear()
                             Navigator.flightRoute.append(PositionProvider.lastValidCoordinate)
                             Navigator.flightRoute.append(waypointDescriptionDialog.waypoint)
                             Global.toast.doToast(qsTr("New flight route: direct to %1.").arg(waypointDescriptionDialog.waypoint.extendedName))
+                            addMenu.close()
+                            waypointDescriptionDialog.close()
                         }
-                        addMenu.close()
                     }
-
                 }
 
                 Rectangle {
@@ -509,7 +512,6 @@ CenteringDialog {
                         // Mention Object to ensure that property gets updated
                         // when flight route changes
                         Navigator.flightRoute.size
-
                         return Navigator.flightRoute.canAppend(waypointDescriptionDialog.waypoint)
                     }
 
@@ -517,6 +519,7 @@ CenteringDialog {
                         PlatformAdaptor.vibrateBrief()
                         Navigator.flightRoute.append(waypointDescriptionDialog.waypoint)
                         addMenu.close()
+                        waypointDescriptionDialog.close()
                         Global.toast.doToast(qsTr("Added %1 to route.").arg(waypointDescriptionDialog.waypoint.extendedName))
                     }
                 }
@@ -535,6 +538,7 @@ CenteringDialog {
                         PlatformAdaptor.vibrateBrief()
                         Navigator.flightRoute.insert(waypointDescriptionDialog.waypoint)
                         addMenu.close()
+                        waypointDescriptionDialog.close()
                         Global.toast.doToast(qsTr("Inserted %1 into route.").arg(waypointDescriptionDialog.waypoint.extendedName))
                     }
                 }
@@ -550,12 +554,13 @@ CenteringDialog {
                         return Navigator.flightRoute.contains(waypointDescriptionDialog.waypoint)
                     }
                     onTriggered: {
-                        PlatformAdaptor.vibrateBrief()
-                        addMenu.close()
+                        PlatformAdaptor.vibrateBrief()                        
                         var index = Navigator.flightRoute.lastIndexOf(waypointDescriptionDialog.waypoint)
                         if (index < 0)
                             return
                         Navigator.flightRoute.removeWaypoint(index)
+                        addMenu.close()
+                        waypointDescriptionDialog.close()
                         Global.toast.doToast(qsTr("Removed %1 from route.").arg(waypointDescriptionDialog.waypoint.extendedName))
                     }
                 }
@@ -698,12 +703,13 @@ CenteringDialog {
             PlatformAdaptor.vibrateBrief()
             Navigator.flightRoute.clear()
             Navigator.flightRoute.append(waypointDescriptionDialog.waypoint)
-            close()
+            overwriteDialog.close()
+            waypointDescriptionDialog.close()
             Global.toast.doToast(qsTr("New flight route: direct to %1.").arg(waypointDescriptionDialog.waypoint.extendedName))
         }
         onRejected: {
             PlatformAdaptor.vibrateBrief()
-            close()
+            overwriteDialog.close()
             waypointDescriptionDialog.open()
         }
     }
@@ -718,6 +724,7 @@ CenteringDialog {
             newWP.notes = newNotes
             newWP.coordinate = QtPositioning.coordinate(newLatitude, newLongitude, newAltitudeMeter)
             WaypointLibrary.replace(waypointDescriptionDialog.waypoint, newWP)
+            waypointDescriptionDialog.close()
             Global.toast.doToast(qsTr("Modified entry %1 in library.").arg(newWP.extendedName))
         }
     }
@@ -734,6 +741,7 @@ CenteringDialog {
             newWP.notes = newNotes
             newWP.coordinate = QtPositioning.coordinate(newLatitude, newLongitude, newAltitudeMeter)
             WaypointLibrary.add(newWP)
+            waypointDescriptionDialog.close()
             Global.toast.doToast(qsTr("Added %1 to waypoint library.").arg(newWP.extendedName))
         }
     }
@@ -751,6 +759,7 @@ CenteringDialog {
         onAccepted: {
             PlatformAdaptor.vibrateBrief()
             WaypointLibrary.remove(removeDialog.waypoint)
+            waypointDescriptionDialog.close()
             Global.toast.doToast(qsTr("Waypoint removed from device"))
         }
         onRejected: {